Refactor API response types to api-helper
