+++ src/library/components/PostCodeAddressSearch/PostCodeAddressSearch.types.ts
@@ -1,4 +1,4 @@
-import { SignpostLinkProp } from "../../../structure/SignpostLinks/SignpostLinks.types"
+import { SignpostLinkProp } from '../../structure/SignpostLinks/SignpostLinks.types';
 
 export interface PostCodeAddressSearchProps {
   /**
@@ -10,15 +10,15 @@ export interface PostCodeAddressSearchProps {
    */
   formError?: boolean;
   /**
-  * An array of the SignpostLinks
-  */
+   * An array of the SignpostLinks
+   */
   signPostLinks?: Array<SignpostLinkProp>;
   /**
-  * Optional, but recommended - a link to send the user to the other council's (West or North) specific service page, defaults to the homepage
-  */
+   * Optional, but recommended - a link to send the user to the other council's (West or North) specific service page, defaults to the homepage
+   */
   otherCouncilLink?: string;
   /**
-  * Optional - changes it to only give back the area you are from
-  */
+   * Optional - changes it to only give back the area you are from
+   */
   isUnitary?: boolean;
 }
